Only create session after clicking pay button
